+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ Features:
 
 - *Backwards-incompatible*: ``fields.Field._deserialize`` now takes ``attr`` and ``data`` as arguments (:issue:`172`). Thanks :user:`alexmic` and :user:`kevinastone` for the suggestion.
 - Allow a ``Field's`` ``attribute`` to be modified during deserialization (:issue:`266`). Thanks :user:`floqqi`.
-- Allow partially-valid data to be return for ``Nested`` fields (:issue:`269`). Thanks :user:`jomag` for the suggestion.
+- Allow partially-valid data to be returned for ``Nested`` fields (:issue:`269`). Thanks :user:`jomag` for the suggestion.
 - Add ``Schema.on_bind_field`` hook which allows a ``Schema`` to modify its fields when they are bound.
 - Stricter validation of string, boolean, and number fields (:issue:`231`). Thanks :user:`touilleMan` for the suggestion.
 - Improve consistency of error messages.